Removed requests page and moved sign controller to top level
