+++ app/views/layouts/admin.html.haml
@@ -61,11 +61,6 @@
                     %i.fi-wrench
                     Settings
               %li
-                = link_to admin_requests_path do
-                  %span.admin-dashboard-sidebar-text
-                    %i.fi-list-bullet
-                    Signbank Logs
-              %li
                 = link_to edit_admin_user_path do
                   %span.admin-dashboard-sidebar-text
                     %i.fi-torsos-female-male
